requirements.txt
pytorch
torchvision
jupyter
